#16 code review
