+++ src/game/cell.js
@@ -7,10 +7,6 @@ class Cell {
     this.type = type;
     this.isHit = false;
   }
-
-  changeType() {
-    this.type = this.type + '-hit';
-  }
 }
 
 export default Cell;
